Insert/copy and edit all components now for both inbound and outbound
