+++ backend/server.js
@@ -344,9 +344,9 @@ const options = {
   ca: fs.readFileSync('../../../../WebApp_Cert/NewWebAppChain.pem')
 };
 
-// https.createServer(options, frontend).listen(SPA_PORT, () => {
-//   console.log(`✅ Frontend (build) served at https://localhost:${SPA_PORT}`);
-// });
+https.createServer(options, frontend).listen(SPA_PORT, () => {
+  console.log(`✅ Frontend (build) served at https://localhost:${SPA_PORT}`);
+});
 
 https.createServer(options, app).listen(port, () => {
   console.log(`✅ Server running at https://localhost:${port}`);
