+++ frontend/src/pages/translations/components/outbound_rule_change.js
@@ -1,388 +1,26 @@
-// Import necessary modules and components
-import React, { useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
 import Select from 'react-select';
+const OutboundRuleChange = ({ 
+    existingRules, 
+    handleRuleChange, 
+    showExisting,
+    form,
+    handleSubmit,
+    handleBack,
+    isEditMode,
+    openHelp,
+    handleSelectChange,
+    handleInputChange,
+    tableOptions,
+    fieldOptions,
+    outputTypeValues,
+    rules,
+    operators,
+    handleRemoveRule,
+    handleAddRule
+}) => {
 
-const TranslationTableRulesOutbound = () => {
-    const location = useLocation();
-    const navigate = useNavigate();
-
-    // Help document URL and opener
-    const HELP_URL = '/help/HelpTranslations.pdf';
-    const openHelp = () => {
-        window.open(HELP_URL, '_blank', 'noopener,noreferrer');
-    };
-
-    // Check if we're in edit mode
-    const [isEditMode, setIsEditMode] = useState(false);
-    const [originalSeq, setOriginalSeq] = useState('');
-    const [originalCustomerNo, setOriginalCustomerNo] = useState(''); // renamed from originalEndDate
-    const [originalTable, setOriginalTable] = useState('');
-    const [originalField, setOriginalField] = useState('');
-    const [prevField, setPrevField] = useState([]);
-    const [prevTable, setPrevTable] = useState([]);
-
-    // Initial state for the form fields
-    const [form, setForm] = useState({
-        trns_trns_tbl: '',
-        trns_trns_fld: '',
-        trns_seq: '',
-        trns_customer_no: '', // replaced dates with customer no
-        trns_output_value: '',
-        trns_output_type: '',
-    });
-
-    // On mount, parse query params and set defaults
-    useEffect(() => {
-        const params = new URLSearchParams(location.search);
-        const mode = params.get('mode');
-        const table = params.get('table') || '';
-        const field = params.get('field') || '';
-        const seq = params.get('seq') || '';
-        const customerNo = params.get('cust_no') || ''; // replaced start/end
-        const outputType = params.get('outputType') || '';
-        const sourceComp = params.get('sourceComp') || '';
-        const operator = params.get('operator') || '';
-        const value = params.get('value') || '';
-        const outputValue = params.get('outputValue') || '';
-        const prevTable = params.get('searchTable') || '';
-        const prevField = params.get('searchField') || '';
-
-
-        setPrevTable(prevTable);
-        setPrevField(prevField);
-        // Set edit mode
-        if (mode === 'edit' || mode === 'copy') {
-            mode === 'edit' ? setIsEditMode(true) : null
-            setOriginalSeq(seq);
-            if (mode === 'edit') {
-                setOriginalCustomerNo(customerNo);
-                setOriginalTable(table);
-                setOriginalField(field);
-            }
-
-            
-            // Parse rule arrays from URL params
-            let parsedRules = [{ comp: '', operator: '', value: '' }];
-            
-            if (sourceComp || operator || value) {
-                const sourceCompArray = sourceComp ? sourceComp.split(',') : [''];
-                let operatorArray = operator ? operator.split(',') : [''];
-                let valueArray;
-                
-                // Try to parse value as JSON (for complex arrays), otherwise split by comma
-                try {
-                    const parsedValue = JSON.parse(value);
-                    if (Array.isArray(parsedValue)) {
-                        valueArray = parsedValue.map(v => Array.isArray(v) ? v.join(',') : v);
-                    } else {
-                        valueArray = [parsedValue];
-                    }
-                } catch {
-                    valueArray = value ? value.split(',') : [''];
-                }
-
-                // Create rules array with matching lengths
-                const maxLength = Math.max(sourceCompArray.length, operatorArray.length, valueArray.length);
-                parsedRules = [];
-                for (let i = 0; i < maxLength; i++) {
-                    parsedRules.push({
-                        comp: sourceCompArray[i] || '',
-                        operator: operatorArray[i] || '',
-                        value: valueArray[i] || ''
-                    });
-                }
-            }
-
-            setRules(parsedRules);
-        }
-
-        setForm(prev => ({
-            ...prev,
-            trns_trns_tbl: mode === 'copy' ? '' : table,
-            trns_trns_fld: mode === 'copy' ? '' : field,
-            trns_seq: mode === 'copy' ? '' : seq,
-            trns_customer_no: customerNo, // set from query
-            trns_output_type: outputType,
-            trns_output_value: outputValue
-        }));
-
-
-    }, [location.search]);
-
-    // Rule rows: each is {comp, operator, value}
-    const [rules, setRules] = useState([
-        { comp: '', operator: '', value: '' }
-    ]);
-
-    // Output type Values (display key and value for DB)
-    const [outputTypeValues, setOutputTypeValues] = useState([
-        { key: 'Character', value: 'Character' },
-        { key: 'Numeric', value: 'Numeric' },
-        { key: 'Add Row', value: 'ADD_ROW' },
-        { key: 'Expression', value: 'Expression' },
-        { key: 'Exclude', value: 'EXCLUDE' }
-    ]);
-
-    const [operators, setOperators] = useState([
-        '=', '<>', 'IN', 'NOT IN', 'IS NULL', 'IS NOT NULL'
-    ]);
-
-    // Table and field options
-    const [tableOptions, setTableOptions] = useState([]);
-    const [fieldOptions, setFieldOptions] = useState([]);
-
-    // Existing translation rules for selected table/field
-    const [existingRules, setExistingRules] = useState([]);
-
-    // Fetch table names on mount
-    useEffect(() => {
-        fetch('https://localhost:5000/TranslationTable/Tables') 
-            .then(res => res.json())
-            .then(data => setTableOptions(data.tables || []))
-            .catch(() => setTableOptions([]));
-            
-    }, []);
-
-
-
-    // Fetch field names when table changes
-    useEffect(() => {
-        if (form.trns_trns_tbl) {
-            fetch(`https://localhost:5000/TranslationTable/Tables/${encodeURIComponent(form.trns_trns_tbl)}/Fields`)
-                .then(res => res.json())
-                .then(data => {
-                    const fields = data.fields || [];
-                    setFieldOptions(fields);
-                
-                })
-                .catch(() => {
-                    setFieldOptions([]);
-                });
-        } else {
-            setFieldOptions([]);
-        }
-    }, [form.trns_trns_tbl]);
-
-    // Fetch existing translation rules when table, field, and customer no change (only in insert mode)
-    useEffect(() => {
-        const hasAll = !!form.trns_trns_tbl && !!form.trns_trns_fld && !!form.trns_customer_no;      
-        if (!isEditMode && hasAll) {
-            const url = `https://localhost:5000/TranslationTable/RulesOutbound?table=${encodeURIComponent(form.trns_trns_tbl)}&field=${encodeURIComponent(form.trns_trns_fld)}&customerNo=${encodeURIComponent(form.trns_customer_no)}`;
-            fetch(url)
-                .then(res => res.json())
-                .then(data => {
-                    const all = data.rules || [];
-                    // Prefer backend filter; also safeguard client-side by matching either property name
-                    const customer = (form.trns_customer_no ?? '').toString().trim();
-                    const filtered = all.filter(r => {
-                        const cust = (r.trns_cust_no ?? r.trns_customer_no ?? '').toString().trim();
-                        return cust === customer;
-                    });
-                    setExistingRules(filtered);
-                })
-                .catch(() => setExistingRules([]));
-        } else {
-            setExistingRules([]);
-        }
-    }, [form.trns_trns_tbl, form.trns_trns_fld, form.trns_customer_no, isEditMode]);
-
-
-    // Handle input changes for main fields
-    const handleSelectChange = (field, option) => {
-        setForm(prev => {
-            const next = { ...prev, [field]: option ? option.value : '' };
-            if (field === 'trns_trns_tbl') {
-                // Reset field when table changes
-                next.trns_trns_fld = '';
-            }
-            return next;
-        });
-    };
-
-    const handleInputChange = (e) => {
-        const { name, value } = e.target;
-        setForm(prev => ({ ...prev, [name]: value }));
-    };
-
-    // Handle rule row changes
-    const handleRuleChange = (idx, field, value) => {
-        setRules(prev => prev.map((row, i) => i === idx ? { ...row, [field]: value } : row));
-    };
-
-    // Add a new rule row
-    const handleAddRule = () => {
-        setRules(prev => [...prev, { comp: '', operator: '', value: '' }]);
-    };
-
-    // Remove a rule row
-    const handleRemoveRule = (idx) => {
-        setRules(prev => prev.length > 1 ? prev.filter((_, i) => i !== idx) : prev);
-    };
-
-    // Handle back navigation
-    const handleBack = () => {
-        navigate(-1);
-    };
-
-
-    // Submit handler
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        // Validate all form fields
-        const requiredFields = [
-            'trns_trns_tbl',
-            'trns_trns_fld',
-            'trns_seq',
-            'trns_customer_no', // now required instead of dates
-            'trns_output_value',
-            'trns_output_type',
-        ];
-        for (let field of requiredFields) {
-            if (!form[field] || form[field].toString().trim() === '') {
-                alert(`Please fill field ${field}.`);
-                return;
-            }
-        }
-        // Validate all rule rows
-        for (let i = 0; i < rules.length; i++) {
-            const row = rules[i];
-            if (!row.comp.toString().trim() || !row.operator.toString().trim() || !row.value.toString().trim()) {
-                alert('Please fill out all rule comparison fields.');
-                return;
-            }
-        }
-
-        // Normalize value: if operator is IN/NOT IN, accept "1, 2" and wrap into "{1,2}"; also convert [a,b] -> {a,b}
-        const normalizeValue = (op, val) => {
-            const s = (val ?? '').toString().trim();
-            const upper = (op ?? '').toString().toUpperCase();
-            if (upper === 'IN' || upper === 'NOT IN') {
-                if (s.startsWith('{') && s.endsWith('}')) return s; // already brace format
-                if (s.startsWith('[') && s.endsWith(']')) return `{${s.slice(1, -1)}}`;
-                const items = s.split(',').map(x => x.trim()).filter(Boolean);
-                return `{${items.join(',')}}`;
-            }
-            return s;
-        };
-
-        const normalizedRules = rules.map(r => ({
-            ...r,
-            value: normalizeValue(r.operator, r.value)
-        }));
-
-        // Helper: format one element for Postgres array literal
-        const formatPgArrayElem = (v) => {
-            if (v === null || v === undefined) return 'NULL';
-            const s = String(v);
-            // Escape backslashes and quotes, then wrap in double quotes
-            const esc = s.replace(/\\/g, '\\\\').replace(/"/g, '\\"');
-            return `"${esc}"`;
-        };
-        // Helper to format JS array as Postgres array string (with proper quoting)
-        const toPgArray = (arr) => '{' + arr.map(formatPgArrayElem).join(',') + '}';
-
-        // Build arrays from rules, format as Postgres arrays
-        const trns_source_comp = toPgArray(rules.map(r => r.comp));
-        const trns_operatione = toPgArray(rules.map(r => r.operator));
-        // Keep each value as a single string element (brace-wrapped for IN/NOT IN)
-        const trns_value = toPgArray(normalizedRules.map(r => r.value));
-
-        // Set audit fields automatically
-        const now = new Date();
-        const pad = n => n.toString().padStart(2, '0');
-        const ymd = now.getFullYear().toString() + pad(now.getMonth() + 1) + pad(now.getDate());
-        const hms = pad(now.getHours()) + pad(now.getMinutes()) + pad(now.getSeconds());
-
-        // Build payload for backend
-        const payload = {
-            trns_trns_tbl: form.trns_trns_tbl,
-            trns_trns_fld: form.trns_trns_fld,
-            trns_seq: form.trns_seq,
-            trns_customer_no: form.trns_customer_no, // include customer no
-            trns_source_comp,
-            trns_operatione,
-            trns_value,
-            trns_output_value: form.trns_output_value,
-            trns_output_type: form.trns_output_type,
-            trns_crt_dte: ymd,
-            trns_crt_tme: hms,
-        };
-
-        // Add original identifiers for updates
-        if (isEditMode) {
-            payload.original_seq = originalSeq;
-            payload.original_trns_trns_tbl = originalTable;
-            payload.original_trns_trns_fld = originalField;
-            payload.original_customer_no = originalCustomerNo;
-        }
-
-        // Choose endpoint and method based on mode
-        const endpoint = isEditMode 
-            ? 'https://localhost:5000/TranslationTable/UpdateRuleOutbound' 
-            : 'https://localhost:5000/TranslationTable/NewRuleOutbound';
-        const method = isEditMode ? 'PUT' : 'POST';
-
-        fetch(endpoint, {
-            method,
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(payload)
-        })
-        .then(async res => {
-            let data;
-            try {
-                data = await res.json();
-            } catch {
-                data = {};
-            }
-            const successMessage = isEditMode ? 'Rule Updated' : 'Rule Added';
-            const successCheck = isEditMode ? 
-                (data.message && data.message.includes('Rule Updated')) :
-                (data.message && data.message.includes('Rule Added'));
-                
-            if (res.ok && successCheck) {
-                alert(successMessage);
-                // Clear all fields only in insert mode
-                if (!isEditMode) {
-                    setForm({
-                        trns_trns_tbl: '',
-                        trns_trns_fld: '',
-                        trns_seq: '',
-                        trns_customer_no: '', // reset
-                        trns_output_value: '',
-                        trns_output_type: '',
-                    });
-                    setRules([{ comp: '', operator: '', value: '' }]);
-                }
-            } else {
-                const errorMessage = isEditMode ? 'Failed to update rule' : 'Failed to add rule';
-                alert((data && data.error) ? data.error : errorMessage);
-            }
-        })
-        .catch(err => {
-            const errorMessage = isEditMode ? 'Failed to update rule' : 'Failed to add rule';
-            alert(errorMessage);
-            console.error('Error submitting rule:', err);
-        });
-    };
-
-    // Layout: flex row if existing rules, else center
-    const showExisting = existingRules.length > 0 && !!form.trns_trns_tbl && !!form.trns_trns_fld && !!form.trns_customer_no;
-    return (
-        <div
-            style={{
-                width: '100%',
-                zIndex: 5,
-                display: 'flex',
-                flexDirection: showExisting ? 'row' : 'column',
-                justifyContent: showExisting ? 'center' : 'center',
-                alignItems: showExisting ? 'flex-start' : 'center',
-                gap: showExisting ? 40 : 0,
-                marginTop: 32,
-                minHeight: '80vh',
-            }}
-        >
-            {/* Existing rules table centered with form, 10px gap */}
+return(
+    <>
             {showExisting && (
                 <div style={{background: '#fafafa', borderRadius: 8, boxShadow: '0 2px 8px #eee', padding: 16, minWidth: 350, width: 'max-content', flexShrink: 0, margin: 0}}>
                     <h3>Existing Translation Rules</h3>
@@ -566,14 +204,14 @@ const TranslationTableRulesOutbound = () => {
                         {/* Customer No and Sequence on the same line */}
                         <div style={{ display: 'flex', gap: 12, width: '100%' }}>
                             <div style={{ flex: '1 1 50%', display: 'flex', flexDirection: 'column', minWidth: 0 }}>
-                                <label htmlFor="trns_customer_no" style={{ marginBottom: 2, fontWeight: 500 }}>
+                                <label htmlFor="trns_cust_no" style={{ marginBottom: 2, fontWeight: 500 }}>
                                     Customer No<span style={{ color: 'red' }}>*</span>
                                 </label>
                                 <input
-                                    id="trns_customer_no"
-                                    name="trns_customer_no"
+                                    id="trns_cust_no"
+                                    name="trns_cust_no"
                                     type="text"
-                                    value={form.trns_customer_no}
+                                    value={form.trns_cust_no}
                                     onChange={handleInputChange}
                                     required
                                     style={{ width: '100%' }}
@@ -665,9 +303,8 @@ const TranslationTableRulesOutbound = () => {
                     <button type="submit" style={{marginTop: 8, width: '100%', padding: 12, background: '#1976d2', color: '#fff', border: 'none', borderRadius: 4, fontSize: 16}}>{isEditMode ? 'Update Rule' : 'Insert Rule'}</button>
                 </form>
             </div>
-        </div>
-        
-    );
-};
+            </>
+        );
+    };
 
-export default TranslationTableRulesOutbound;
+    export default OutboundRuleChange;
