+++ frontend/src/pages/translations/translationtablerules.js
@@ -1,11 +1,44 @@
 // Import necessary modules and components
 import React, { useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import Select from 'react-select';
+
+import OutboundRuleChange from "./components/outbound_rule_change";
+import InboundRuleChange from "./components/inbound_rule_change";
+const mode = 'I'; // or get this from somewhere - you need to define mode
 
 const TranslationTableRules = () => {
     const location = useLocation();
     const navigate = useNavigate();
+    
+    // Get mode from URL parameters
+    const urlParams = new URLSearchParams(location.search);
+    const mode = urlParams.get('mode') || 'I'; // Default to 'I' if not specified
+    
+    // Initialize column filters based on mode
+    const [columnFilters, setColumnFilters] = useState(() => {
+        if (mode === 'I') {
+            return {
+                seq: '',
+                table: '',
+                field: '',
+                sourceComp: '',
+                operator: '',
+                value: '',
+                outputValue: ''
+            };
+        } else {
+            return {
+                seq: '',
+                table: '',
+                field: '',
+                customerNo: '', // Use customerNo, not cust_no
+                sourceComp: '',
+                operator: '',
+                value: '',
+                outputValue: ''
+            };
+        }
+    });
 
     // Help document URL and opener
     const HELP_URL = '/help/HelpTranslations.pdf';
@@ -17,13 +50,12 @@ const TranslationTableRules = () => {
     const [isEditMode, setIsEditMode] = useState(false);
     const [originalSeq, setOriginalSeq] = useState('');
     const [originalEndDate, setOriginalEndDate] = useState('');
+    const [originalCustomerNo, setOriginalCustomerNo] = useState('');
     const [originalTable, setOriginalTable] = useState('');
     const [originalField, setOriginalField] = useState('');
-    const [prevField, setPrevField] = useState([]);
-    const [prevTable, setPrevTable] = useState([]);
 
     // Initial state for the form fields
-    const [form, setForm] = useState({
+    const [form, setForm] = useState(mode === 'I' ?{
         trns_trns_tbl: '',
         trns_trns_fld: '',
         trns_end_dte: '',
@@ -31,7 +63,15 @@ const TranslationTableRules = () => {
         trns_strt_dte: '',
         trns_output_value: '',
         trns_output_type: '',
-    });
+    }:{
+        trns_trns_tbl: '',
+        trns_trns_fld: '',
+        trns_seq: '',
+        trns_cust_no: '', 
+        trns_output_value: '',
+        trns_output_type: '',
+    }
+);
 
     // On mount, parse query params and set defaults
     useEffect(() => {
@@ -49,16 +89,15 @@ const TranslationTableRules = () => {
         const outputValue = params.get('outputValue') || '';
         const prevTable = params.get('searchTable') || '';
         const prevField = params.get('searchField') || '';
+        const prevCustNo = params.get('cust_no') || '';
 
 
-        console.log(prevField, prevTable);
-        setPrevTable(prevTable);
-        setPrevField(prevField);
         // Set edit mode
         if (mode === 'edit' || mode === 'copy') {
             mode === 'edit' ? setIsEditMode(true) : null
             setOriginalSeq(seq);
             setOriginalEndDate(endDate);
+            mode === 'edit' || mode === 'copy' ? setOriginalCustomerNo(prevCustNo) : null;
             mode === 'edit' ? setOriginalTable(table) : null;
             mode === 'edit' ? setOriginalField(field) : null;
 
@@ -98,13 +137,21 @@ const TranslationTableRules = () => {
             setRules(parsedRules);
         }
 
-        setForm(prev => ({
+        setForm(prev => (mode === 'I' ? {
             ...prev,
             trns_trns_tbl: mode === 'copy' ? '' : table,
             trns_trns_fld: mode === 'copy' ? '' : field,
             trns_seq: mode === 'copy' ? '' : seq,
             trns_strt_dte: startDate,
             trns_end_dte: endDate,
+            trns_output_type: outputType,
+            trns_output_value: outputValue
+        } :{
+            ...prev,
+            trns_trns_tbl: mode === 'copy' ? '' : table,
+            trns_trns_fld: mode === 'copy' ? '' : field,
+            trns_seq: mode === 'copy' ? '' : seq,
+            trns_cust_no: prevCustNo, 
             trns_output_type: outputType,
             trns_output_value: outputValue
         }));
@@ -169,7 +216,9 @@ const TranslationTableRules = () => {
     // Fetch existing translation rules when table or field changes (only in insert mode)
     useEffect(() => {
         if (!isEditMode && form.trns_trns_tbl && form.trns_trns_fld) {
-            fetch(`https://az-cld-ivap-d1:5000/TranslationTable/Rules?table=${encodeURIComponent(form.trns_trns_tbl)}&field=${encodeURIComponent(form.trns_trns_fld)}`)
+            fetch( mode === 'I' ?
+                `https://az-cld-ivap-d1:5000/TranslationTable/Rules?table=${encodeURIComponent(form.trns_trns_tbl)}&field=${encodeURIComponent(form.trns_trns_fld)}`
+                : `https://az-cld-ivap-d1:5000/TranslationTable/RulesOutbound?table=${encodeURIComponent(form.trns_trns_tbl)}&field=${encodeURIComponent(form.trns_trns_fld)}}`)
                 .then(res => res.json())
                 .then(data => setExistingRules(data.rules || []))
                 .catch(() => setExistingRules([]));
@@ -221,7 +270,7 @@ const TranslationTableRules = () => {
     const handleSubmit = (e) => {
         e.preventDefault();
         // Validate all form fields
-        const requiredFields = [
+        const requiredFields = mode === 'I' ? [
             'trns_trns_tbl',
             'trns_trns_fld',
             'trns_end_dte',
@@ -229,7 +278,15 @@ const TranslationTableRules = () => {
             'trns_strt_dte',
             'trns_output_value',
             'trns_output_type',
+        ] : [
+            'trns_trns_tbl',
+            'trns_trns_fld',
+            'trns_seq',
+            'trns_cust_no', 
+            'trns_output_value',
+            'trns_output_type',
         ];
+
         for (let field of requiredFields) {
             if (!form[field] || form[field].toString().trim() === '') {
                 alert(`Please fill field ${field}.`);
@@ -287,7 +344,7 @@ const TranslationTableRules = () => {
         const hms = pad(now.getHours()) + pad(now.getMinutes()) + pad(now.getSeconds());
 
         // Build payload for backend
-        const payload = {
+        const payload = mode === 'I' ?{
             trns_trns_tbl: form.trns_trns_tbl,
             trns_trns_fld: form.trns_trns_fld,
             trns_end_dte: form.trns_end_dte,
@@ -300,21 +357,38 @@ const TranslationTableRules = () => {
             trns_output_type: form.trns_output_type,
             trns_crt_dte: ymd,
             trns_crt_tme: hms,
-            // Add other audit/user fields as needed
+        } : {
+            trns_trns_tbl: form.trns_trns_tbl,
+            trns_trns_fld: form.trns_trns_fld,
+            trns_seq: form.trns_seq,
+            trns_cust_no: form.trns_cust_no, 
+            trns_source_comp,
+            trns_operatione,
+            trns_value,
+            trns_output_value: form.trns_output_value,
+            trns_output_type: form.trns_output_type,
+            trns_crt_dte: ymd,
+            trns_crt_tme: hms,
         };
 
         // Add original sequence for updates
-        if (isEditMode) {
+        if (isEditMode && mode === 'I') {
             payload.original_seq = originalSeq;
             payload.original_end_dte = originalEndDate;
             payload.original_trns_trns_tbl = originalTable;
             payload.original_trns_trns_fld = originalField;
         }
+        else if (isEditMode) {
+            payload.original_seq = originalSeq;
+            payload.original_trns_trns_tbl = originalTable;
+            payload.original_trns_trns_fld = originalField;
+            payload.original_cust_no = originalCustomerNo;
+        }
 
         // Choose endpoint and method based on mode
         const endpoint = isEditMode 
-            ? 'https://az-cld-ivap-d1:5000/TranslationTable/UpdateRule' 
-            : 'https://az-cld-ivap-d1:5000/TranslationTable/NewRule';
+            ? (mode === 'I' ? 'https://az-cld-ivap-d1:5000/TranslationTable/UpdateRule' : 'https://az-cld-ivap-d1:5000/TranslationTable/UpdateRuleOutbound')
+            : (mode === 'I' ? 'https://az-cld-ivap-d1:5000/TranslationTable/NewRule' : 'https://az-cld-ivap-d1:5000/TranslationTable/NewRuleOutbound');
         const method = isEditMode ? 'PUT' : 'POST';
 
         fetch(endpoint, {
@@ -338,12 +412,19 @@ const TranslationTableRules = () => {
                 alert(successMessage);
                 // Clear all fields only in insert mode
                 if (!isEditMode) {
-                    setForm({
+                    setForm( mode === 'I' ? {
                         trns_trns_tbl: '',
                         trns_trns_fld: '',
                         trns_end_dte: '',
                         trns_seq: '',
                         trns_strt_dte: '',
+                        trns_output_value: '',
+                        trns_output_type: '',
+                    } : {
+                        trns_trns_tbl: '',
+                        trns_trns_fld: '',
+                        trns_seq: '',
+                        trns_cust_no: '', 
                         trns_output_value: '',
                         trns_output_type: '',
                     });
@@ -377,258 +458,49 @@ const TranslationTableRules = () => {
                 minHeight: '80vh',
             }}
         >
-            {/* Existing rules table centered with form, 10px gap */}
-            {showExisting && (
-                <div style={{background: '#fafafa', borderRadius: 8, boxShadow: '0 2px 8px #eee', padding: 16, minWidth: 350, width: 'max-content', flexShrink: 0, margin: 0}}>
-                    <h3>Existing Translation Rules</h3>
-                    <table style={{width: '100%', borderCollapse: 'collapse'}}>
-                        <thead>
-                            <tr style={{background: '#f0f0f0'}}>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Seq</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Source Comp</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Operator</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Value</th>    
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Output Value</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {existingRules.map((rule, i) => (
-                                <tr key={i}>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>{rule.trns_seq}</td>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        {Array.isArray(rule.trns_source_comp)
-                                            ? rule.trns_source_comp.map((v, idx) => (
-                                                <div key={idx}>{v}</div>
-                                            ))
-                                            : rule.trns_source_comp}
-                                    </td>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        {Array.isArray(rule.trns_operatione)
-                                            ? rule.trns_operatione.map((v, idx) => (
-                                                <div key={idx}>{v}</div>
-                                            ))
-                                            : rule.trns_operatione}
-                                    </td>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        {(() => {
-                                            const prettyBraceString = (s) => {
-                                                if (typeof s === 'string') {
-                                                    const t = s.trim();
-                                                    if (t.startsWith('{') && t.endsWith('}')) {
-                                                        return `[${t.slice(1, -1)}]`;
-                                                    }
-                                                }
-                                                return s;
-                                            };
-
-                                            const renderVal = (v, i) => {
-                                                if (Array.isArray(v)) {
-                                                    return `[${v.join(',')}]`;
-                                                }
-                                                return <div key={i}>{prettyBraceString(v)}</div>;
-                                            };
-
-                                            const val = rule.trns_value;
-                                            if (Array.isArray(val)) {
-                                                // If it's a single nested array (e.g. [[1,2,3]]), show as [1,2,3]
-                                                if (val.length === 1 && Array.isArray(val[0])) {
-                                                    return `[${val[0].join(',')}]`;
-                                                }
-                                                // If it's a flat array (no nested arrays), stack vertically and pretty print brace strings
-                                                if (val.length > 1 && !val.some(Array.isArray)) {
-                                                    return val.map((v, i) => renderVal(v, i));
-                                                }
-                                                // If it's a mixed array or nested arrays, pretty print nested arrays
-                                                if (val.some(Array.isArray)) {
-                                                    return val.map((v, i) => Array.isArray(v) ? `[${v.join(',')}]` : renderVal(v, i));
-                                                }
-                                                // Single value array
-                                                if (val.length === 1) {
-                                                    return prettyBraceString(val[0]);
-                                                }
-                                                // Empty array
-                                                return '';
-                                            } else {
-                                                return prettyBraceString(val);
-                                            }
-                                        })()}
-                                    </td>
-                                    <td style={{ padding: 4, border: '1px solid #ccc', whiteSpace: 'pre-wrap', wordBreak: 'break-word', overflowWrap: 'anywhere' }}>{rule.trns_output_value}</td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
-                </div>
-            )}
-            {/* Centered form, with 10px gap if existing rules */}
-            <div
-                style={{
-                    display: 'flex',
-                    justifyContent: showExisting ? 'center' : 'center',
-                    alignItems: showExisting ? 'flex-start' : 'center',
-                    minWidth: 0,
-                    margin: 0,
-                    width: showExisting ? undefined : '100%',
-                }}
-            >
-                <form onSubmit={handleSubmit} style={{width: '600px', background: '#fff', padding: 16, borderRadius: 8, boxShadow: '0 2px 8px #ccc'}}>
-                    <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: 16}}>
-                        <button 
-                            type="button" 
-                            onClick={handleBack}
-                            style={{
-                                background: '#666',
-                                color: '#fff',
-                                border: 'none',
-                                borderRadius: 4,
-                                padding: '8px 16px',
-                                fontSize: 14,
-                                cursor: 'pointer',
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: 4
-                            }}
-                        >
-                            ← Back
-                        </button>
-                        <h2 style={{margin: 0}}>{isEditMode ? 'Edit Translation Table Rule' : 'Insert Translation Table Rule'}</h2>
-                        <button
-                            type="button"
-                            onClick={openHelp}
-                            title="Help"
-                            aria-label="Help"
-                            style={{
-                                background: '#1976d2',
-                                color: '#fff',
-                                border: 'none',
-                                borderRadius: '50%',
-                                width: 32,
-                                height: 32,
-                                fontSize: 18,
-                                fontWeight: 700,
-                                cursor: 'pointer',
-                                display: 'flex',
-                                alignItems: 'center',
-                                justifyContent: 'center',
-                                boxShadow: '0 1px 4px #bbb'
-                            }}
-                        >
-                            ?
-                        </button>
-                    </div>
-                    <div style={{display: 'flex', flexWrap: 'wrap', gap: 12}}>
-                        <Select
-                            placeholder={<div>Select a table...</div>}
-                            onChange={(opt) => handleSelectChange('trns_trns_tbl', opt)}
-                            value={form.trns_trns_tbl ? { value: form.trns_trns_tbl, label: form.trns_trns_tbl } : null}
-                            required
-                            styles={{
-                                container: (base) => ({
-                                    ...base,
-                                    width: 'calc((100% - 12px) / 2)',
-                                    flex: '0 0 calc((100% - 12px) / 2)'
-                                }),
-                                control: (base) => ({
-                                    ...base,
-                                    width: '100%'
-                                })
-                            }}
-                            options={tableOptions.map(tbl => ({ value: tbl, label: tbl }))}
-                        />
-                        <Select
-                            placeholder={<div>Select a field...</div>}
-                            name="trns_trns_fld"
-                            onChange={(opt) => handleSelectChange('trns_trns_fld', opt)}
-                            value={form.trns_trns_fld ? { value: form.trns_trns_fld, label: form.trns_trns_fld } : null}
-                            required
-                            styles={{
-                                container: (base) => ({
-                                    ...base,
-                                    width: 'calc((100% - 12px) / 2)',
-                                    flex: '0 0 calc((100% - 12px) / 2)'
-                                }),
-                                control: (base) => ({
-                                    ...base,
-                                    width: '100%'
-                                })
-                            }}
-                            options={fieldOptions.map(fld => ({ value: fld, label: fld }))}
-                            isDisabled={!form.trns_trns_tbl}
-                        />
-                        <div style={{flex: '1 1 45%', display: 'flex', flexDirection: 'column'}}>
-                            <label htmlFor="trns_strt_dte" style={{marginBottom: 2, fontWeight: 500}}>Start Date</label>
-                            <input id="trns_strt_dte" name="trns_strt_dte" type="date" value={form.trns_strt_dte} onChange={handleInputChange} style={{width: '100%'}} />
-                        </div>
-                        <div style={{flex: '1 1 45%', display: 'flex', flexDirection: 'column'}}>
-                            <label htmlFor="trns_end_dte" style={{marginBottom: 2, fontWeight: 500}}>End Date<span style={{color: 'red'}}>*</span></label>
-                            <input id="trns_end_dte" name="trns_end_dte" type="date" value={form.trns_end_dte} onChange={handleInputChange} required style={{width: '100%'}} />
-                        </div>
-                        <input name="trns_seq" placeholder="Sequence (number)" value={form.trns_seq} onChange={handleInputChange} required style={{flex: '1 1 45%'}} />
-                        <select name="trns_output_type" value={form.trns_output_type} onChange={handleInputChange} style={{flex: '1 1 45%'}} required>
-                            <option value="" disabled>Output Type</option>
-                            {outputTypeValues.map(opt => (
-                                <option key={opt.value} value={opt.value}>{opt.key}</option>
-                            ))}
-                        </select>
-                        <div style={{flex: '1 1 45%', display: 'flex', flexDirection: 'column'}}>
-                            <label htmlFor="trns_output_value" style={{marginBottom: 2, fontWeight: 500}}>Output Value</label>
-                            <textarea
-                                id="trns_output_value"
-                                name="trns_output_value"
-                                placeholder="Output Value"
-                                value={form.trns_output_value}
-                                onChange={handleInputChange}
-                                style={{width: '100%', minHeight: 32, resize: 'vertical', fontFamily: 'inherit', fontSize: 16, padding: 6, borderRadius: 4, border: '1px solid #ccc'}}
-                                rows={2}
-                            />
-                        </div>
-                    </div>
-                    <h3 style={{marginTop: 24}}>Rule Comparisons</h3>
-                    <table style={{width: '100%', marginBottom: 16, borderCollapse: 'collapse'}}>
-                        <thead>
-                            <tr style={{background: '#f0f0f0'}}>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Comp</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Operator</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Value</th>
-                                <th style={{padding: 8, border: '1px solid #ccc'}}>Action</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {rules.map((row, idx) => (
-                                <tr key={idx}>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        <textarea value={row.comp} onChange={e => handleRuleChange(idx, 'comp', e.target.value)}
-                                            style={{width: '100%', minHeight: 32, resize: 'vertical', fontSize: 16, padding: 6, borderRadius: 4, border: '1px solid #ccc', fontFamily: 'inherit'}}
-                                            rows={1}
-                                        />
-                                    </td>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        <select value={row.operator} onChange={e => handleRuleChange(idx, 'operator', e.target.value)} style={{width: '100%', minHeight: 32, overflowY: 'auto', fontSize: 16, padding: 6, borderRadius: 4, border: '1px solid #ccc'}}>
-                                            <option value="" disabled>Select Operator</option>
-                                            {operators.map(opt => (
-                                                <option key={opt} value={opt}>{opt}</option>
-                                            ))}
-                                        </select>
-                                    </td>
-                                    <td style={{padding: 4, border: '1px solid #ccc'}}>
-                                        <textarea value={row.value} onChange={e => handleRuleChange(idx, 'value', e.target.value)}
-                                            style={{width: '100%', minHeight: 32, resize: 'vertical', fontSize: 16, padding: 6, borderRadius: 4, border: '1px solid #ccc', fontFamily: 'inherit'}}
-                                            rows={1}
-                                        />
-                                    </td>
-                                    <td style={{padding: 4, border: '1px solid #ccc', textAlign: 'center'}}>
-                                        <button type="button" onClick={() => handleRemoveRule(idx)} disabled={rules.length === 1} style={{background: '#e57373', color: '#fff', border: 'none', borderRadius: 4, padding: '4px 8px', cursor: rules.length === 1 ? 'not-allowed' : 'pointer'}}>Remove</button>
-                                    </td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
-                    <button type="button" onClick={handleAddRule} style={{marginBottom: 16, background: '#4caf50', color: '#fff', border: 'none', borderRadius: 4, padding: '8px 16px', fontSize: 14}}>Add Rule</button>
-                    <button type="submit" style={{marginTop: 8, width: '100%', padding: 12, background: '#1976d2', color: '#fff', border: 'none', borderRadius: 4, fontSize: 16}}>{isEditMode ? 'Update Rule' : 'Insert Rule'}</button>
-                </form>
-            </div>
+           {mode === 'I' ? 
+                <InboundRuleChange
+                    existingRules={existingRules}
+                    handleRuleChange={handleRuleChange}
+                    showExisting={existingRules.length > 0}
+                    form={form}
+                    handleSubmit={handleSubmit}
+                    handleBack={handleBack}
+                    isEditMode={isEditMode}
+                    openHelp={openHelp}
+                    handleSelectChange={handleSelectChange}
+                    handleInputChange={handleInputChange}
+                    tableOptions={tableOptions}
+                    fieldOptions={fieldOptions}
+                    outputTypeValues={outputTypeValues}
+                    rules={rules}
+                    operators={operators}
+                    handleRemoveRule={handleRemoveRule}
+                    handleAddRule={handleAddRule}
+                    mode={mode}
+                /> : 
+                <OutboundRuleChange
+                    existingRules={existingRules}
+                    handleRuleChange={handleRuleChange}
+                    showExisting={existingRules.length > 0}
+                    form={form}
+                    handleSubmit={handleSubmit}
+                    handleBack={handleBack}
+                    isEditMode={isEditMode}
+                    openHelp={openHelp}
+                    handleSelectChange={handleSelectChange}
+                    handleInputChange={handleInputChange}
+                    tableOptions={tableOptions}
+                    fieldOptions={fieldOptions}
+                    outputTypeValues={outputTypeValues}
+                    rules={rules}
+                    operators={operators}
+                    handleRemoveRule={handleRemoveRule}
+                    handleAddRule={handleAddRule}
+                    mode={mode}
+                />
+            }
         </div>
-        
     );
 };
